requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.4.1
astroid==2.9.3
autopep8==1.6.0
bcrypt==3.2.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.72.0
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Mako==1.2.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
orjson==3.6.5
passlib==1.7.4
pathspec==0.9.0
pep8==1.7.1
platformdirs==2.4.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pylint==2.12.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.0
starlette==0.17.1
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wrapt==1.13.3
yapf==0.32.0
zipp==3.7.0
